app/requirements.txt
streamlit
pytube
openai
langchain
transformers
youtube-transcript-api
tiktoken
scikit-learn
python-dotenv
